+++ packages/check-types-mini/changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.0] - 2017-06-18
+### Added
+- ✨ Now, the errors which are caused by misconfiguration of the `check-types-mini` itself will reference it as a source of an error. Once this library is configured correctly, then the errors can be personalised as per `opts.msg`.
+
 ## [2.0.0] - 2017-06-12
 ### Changed
 - ✨ BREAKING API CHANGES. Third argument `msg` moved to `opts.msg`. Fourth argument `optsVarName` moved to `opts.optsVarName`. That was the right thing to do. Sorry for any hassle updating.
@@ -45,3 +49,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [1.5.0]: https://github.com/codsen/check-types-mini/compare/v1.4.0...v1.5.0
 [1.6.0]: https://github.com/codsen/check-types-mini/compare/v1.5.0...v1.6.0
 [2.0.0]: https://github.com/codsen/check-types-mini/compare/v1.6.0...v2.0.0
+[2.1.0]: https://github.com/codsen/check-types-mini/compare/v2.0.0...v2.1.0